requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
numpy>=1.26.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.7.0
bandit>=1.7.9
docker>=7.1.0
pytest>=8.2.0
rich>=13.7.0
colorama>=0.4.6
